Turns and input check added
I've added a turn counter instance variable to the Game class, and when 
a turn is run it adds +1 to the variable. At 9, it ends the game, so the 
game will only last as many turns as there are empty squares.

Also made a correct_input_check method in PLayer that will check whether 
the player's selection is available and, if not, will alert the player 
and prompt them to make a new selection.